+++ .gitmodules
@@ -1,3 +1,3 @@
-[submodule "Client/vendor"]
-	path = Client/vendor
+[submodule "client/deps/ViGEm"]
+	path = client/deps/ViGEm
 	url = https://github.com/ViGEm/ViGEmClient.git
