+++ server/sce_sys/livearea/contents/template.xml
